[CORE-6048] restrict based on admin platform
