+++ projects/v3/src/app/services/utils.service.ts
@@ -2,9 +2,7 @@ import { Injectable, Inject } from '@angular/core';
 import { DOCUMENT } from '@angular/common';
 import { Observable, Subject } from 'rxjs';
 import { map, filter } from 'rxjs/operators';
-import { Platform } from '@ionic/angular';
 import { ModalController } from '@ionic/angular';
-import { ApolloService } from '@v3/services/apollo.service';
 import * as _ from 'lodash';
 import * as moment from 'moment';
 import { Colors, BrowserStorageService } from './storage.service';
@@ -37,7 +35,6 @@ export class UtilsService {
 
   constructor(
     @Inject(DOCUMENT) private document: Document,
-    private apolloService: ApolloService,
     private readonly modalController: ModalController,
     private readonly storageService: BrowserStorageService,
   ) {
@@ -221,21 +218,6 @@ export class UtilsService {
   //   }
   //   this.activitySubjects[key].next(value);
   // }
-
-  // need to clear all Subject for cache
-  async clearCache(): Promise<void> {
-    const apolloClient = this.apolloService.getClient();
-    // clear cache before initialised
-    if (apolloClient) {
-      apolloClient.stop();
-      await apolloClient.clearStore();
-    }
-    //   // initialise the Subject for caches
-    //   this.projectSubject.next(null);
-    //   this.each(this.activitySubjects, (subject, key) => {
-    //     this.activitySubjects[key].next(null);
-    //   });
-  }
 
   getCurrentLocation(): Location {
     return this.document.location;
